# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Daycore v2 API
  version: "2.4.0"
  description: >
    Daycore v2 backend (Go). Autonomous AI day-planning + mood companion.
    Anonymous-first: POST /api/session/init issues an httpOnly, signed session
    cookie (dc_sid) that authorizes all data endpoints. Optional email/OAuth
    sign-in sets an httpOnly JWT cookie (dc_auth); logout revokes all issued JWTs
    (server-side token version). Same-origin web frontends MUST send cookies
    (fetch: credentials:"include"). Native apps and separately-deployed frontends
    use header credentials instead (dual-rail, cookies stay supported): the
    signed session token in X-Session-Token (returned by /api/session/init with
    {"tokenInBody":true}) and/or the JWT in Authorization: Bearer (returned by
    login/register with {"tokenInBody":true}). Check GET /api/version (apiVersion
    contract major) before use — matching rules in FRONTEND_HANDOFF.md.
    The browser extension authenticates to /api/import/* with X-Import-Token;
    admin endpoints use X-Admin-Token.
    Errors are {error, message}. The single SSE endpoint is /api/ai/companion —
    see FRONTEND_HANDOFF.md §B for the v2 frame protocol (NOT the legacy [DONE]).
servers:
  - url: http://localhost:8080
    description: Local (Go serves the API; nginx proxies /api in production)
...
security:
  - sessionCookie: []
  - sessionToken: []
  - bearerAuth: []
paths:
...
  /api/v2/session/init:
    post:
      operationId: sessionInit
      tags: [session]
      summary: Create/return the anonymous session; sets dc_sid cookie.
      description: >
        Optional body {"tokenInBody":true} adds "sessionToken" (the same signed
        value as the dc_sid cookie) to the response for clients without a cookie
        jar; send it back as the X-Session-Token header.

        "timezone" is the device's IANA zone. First contact is the cheapest place
        to learn it and the most important one: the cron entries that fire the
        morning brief, the petrify line, the rhythm day key and every AI clock are
        all derived from the session's zone (docs/ARCHITECTURE.md 「每会话时区」).
        Best-effort — an unknown name is ignored rather than refused, and a session
        that never gets one keeps the deployment default (WORKER_DEFAULT_TZ).
      requestBody: { required: false, content: { application/json: { schema: { type: object, properties: { tokenInBody: { type: boolean }, timezone: { type: string, example: America/Chicago } } } } } }
      responses: { '200': { description: "OK (Session; + sessionToken when tokenInBody)", content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } } }
  /api/v2/session/theme:
    post:
      operationId: sessionTheme
      tags: [session]
      summary: Record the current theme FOR THE CALLING FRONTEND'S FAMILY.
      description: >-
        桌面用琉璃、手机用汀 is the ordinary case, so there is no single "current
        theme" for a session — one per family, and the response echoes which one
        was written.

        ⚠️ Two homes, one per family, not two homes for one value: the fallback
        family stays in `sessions.current_theme` (the column the shipping
        frontend reads), every other family lives in the preferences blob. See
        docs/DATA.md.
      parameters:
        - name: X-Frontend-Build
          in: header
          required: false
          schema: { type: string }
          description: >-
            Which build is asking, and therefore which frontend FAMILY's theme
            token space applies. Absent, or a build this deployment has never
            seen, means the built-in fallback family — which is why a frontend
            that never handshakes behaves exactly as it always did.

            ⚠️ NOT a credential and must never become one. It selects a token
            space and nothing else; a caller that lies about it gets somebody
            else's list of variable names, which is public information (it is in
            every manifest and every stylesheet).
      requestBody: { content: { application/json: { schema: { type: object, properties: { theme: { type: string } } } } } }
      responses: { '200': { description: OK } }
  /api/v2/session/settings:
    patch:
      operationId: sessionSettings
      tags: [session]
      summary: Update assistantName / theme / language / personaPrompt.
      description: >
        language is which of the user's two languages they are reading in right
        now — what the home-page switch flips. It must be one of their pair
        (GET /api/session/preferences); the pair itself is changed through
        PATCH /api/session/preferences, not here. A value outside the pair is
        rejected with 400 unsupported_locale. Reads never fail this way: a
        stored value that falls outside the pair resolves to the primary, so
        changing your pair never breaks your settings page.
      requestBody: { content: { application/json: { schema: { type: object, properties: { assistantName: { type: string }, currentTheme: { type: string }, language: { type: string, enum: [zh-CN, en-US] }, personaPrompt: { type: string, maxLength: 2000 } } } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } } }
...
  /api/v2/ai/companion:
    post:
      operationId: companion
      tags: [ai]
      summary: Companion agent over SSE v2 (tool-calls frame protocol).
      description: >
        Returns text/event-stream. Each frame is `data: {json}\n\n` discriminated
        by `type`: delta | reasoning | tool_start | tool_result | decision_card |
        error | done. Comment heartbeats `: ping`. NO `[DONE]` sentinel. See
        FRONTEND_HANDOFF.md §B. With threadId the server loads AND persists thread
        history; client roles are whitelisted to user/assistant.
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { message: { type: string, description: "May be empty when attachmentIds is not — dropping in a photo with no words is a complete thing to say." }, timezone: { type: string }, location: { type: string, maxLength: 128, description: "Where the device is, free text. Fills in or updates a device-detected location; never overwrites one the user set on the settings page. Used only by the briefs — the chat tool takes a location as an argument." }, assistantName: { type: string }, threadId: { type: string }, attachmentIds: { type: array, maxItems: 20, items: { type: string }, description: "Ids from POST /api/files. Must be this session's unsent uploads; they become part of the persisted user message. Files the chosen model cannot read are described to it by name rather than dropped." }, conversationHistory: { type: array, items: { type: object, properties: { role: { type: string, enum: [user, assistant] }, content: { type: string } } } } } } } } }
      responses: { '200': { description: SSE stream (text/event-stream), content: { text/event-stream: { schema: { type: string }, example: 'data: {"type":"delta","text":"好的"}\n\ndata: {"type":"done"}\n\n' } } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: dc_sid }
    authCookie: { type: apiKey, in: cookie, name: dc_auth }
    sessionToken: { type: apiKey, in: header, name: X-Session-Token, description: "Same signed value as the dc_sid cookie, for clients without a cookie jar. Get it from POST /api/session/init {tokenInBody:true}." }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: "Same JWT as the dc_auth cookie. Get it from login/register {tokenInBody:true}; logout revokes all issued JWTs (token-version bump)." }
    importToken: { type: apiKey, in: header, name: X-Import-Token, description: "Session-scoped token for the browser extension's direct push." }
    adminToken: { type: apiKey, in: header, name: X-Admin-Token, description: "Admin endpoints." }
...
  schemas:
...
    Session:
      type: object
      properties:
        id: { type: string }
        userId: { type: string, nullable: true }
        interactionCount: { type: integer }
        signInPrompted: { type: boolean }
        assistantName: { type: string }
        currentTheme: { type: string }
        language: { type: string, enum: [zh-CN, en-US] }
        personaPrompt: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...